install_dependencies.sh: check for dpkg, assume adavance users know what to pre install..
fixes: https://github.com/RobertCNelson/omap-image-builder/issues/136

+++ scripts/install_dependencies.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 #
-# Copyright (c) 2012-2020 Robert Nelson <robertcnelson@gmail.com>
+# Copyright (c) 2012-2021 Robert Nelson <robertcnelson@gmail.com>
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,24 +26,26 @@ minimal_debootstrap="123"
 host_arch="$(uname -m)"
 
 debootstrap_is_installed () {
-	unset deb_pkgs
-	dpkg -l | grep debootstrap >/dev/null || deb_pkgs="${deb_pkgs}debootstrap "
+	if [ -f /usr/bin/dpkg ] ; then
+		unset deb_pkgs
+		dpkg -l | grep debootstrap >/dev/null || deb_pkgs="${deb_pkgs}debootstrap "
 
-	if [ "x${host_arch}" != "xarmv7l" ] ; then
-		if [ "x${host_arch}" != "xaarch64" ] ; then
-			#FIXME:...
-			#echo "QEMU is un-reliable, thus no longer supported... Spend some Money and buy a real ARMHF device to run this script."
-			#FIXME: comment out the next line to use QEMU
-			#exit 2
-			dpkg -l | grep qemu-user-static >/dev/null || deb_pkgs="${deb_pkgs}qemu-user-static "
-			dpkg -l | grep $(dpkg --print-architecture) | grep -v "qemu-" | grep qemu >/dev/null || deb_pkgs="${deb_pkgs}qemu "
+		if [ "x${host_arch}" != "xarmv7l" ] ; then
+			if [ "x${host_arch}" != "xaarch64" ] ; then
+				#FIXME:...
+				#echo "QEMU is un-reliable, thus no longer supported... Spend some Money and buy a real ARMHF device to run this script."
+				#FIXME: comment out the next line to use QEMU
+				#exit 2
+				dpkg -l | grep qemu-user-static >/dev/null || deb_pkgs="${deb_pkgs}qemu-user-static "
+				dpkg -l | grep $(dpkg --print-architecture) | grep -v "qemu-" | grep qemu >/dev/null || deb_pkgs="${deb_pkgs}qemu "
+			fi
 		fi
-	fi
 
-	if [ "${deb_pkgs}" ] ; then
-		echo "Installing: ${deb_pkgs}"
-		sudo apt-get update
-		sudo apt-get -y install ${deb_pkgs}
+		if [ "${deb_pkgs}" ] ; then
+			echo "Installing: ${deb_pkgs}"
+			sudo apt-get update
+			sudo apt-get -y install ${deb_pkgs}
+		fi
 	fi
 }
 
@@ -52,15 +54,19 @@ debootstrap_what_version () {
 	echo "Log: debootstrap version: 1.0.$test_debootstrap"
 }
 
+install_debootstrap () {
+	if [ -f /usr/bin/dpkg ] ; then
+		#if [[ "$test_debootstrap" < "$minimal_debootstrap" ]] ; then
+		#if [ "$test_debootstrap" -lt "$minimal_debootstrap" ] ; then
+		if [ ! "x$test_debootstrap" = "x$minimal_debootstrap" ] ; then
+			echo "Log: Installing minimal debootstrap version: 1.0.${minimal_debootstrap}..."
+			wget https://rcn-ee.com/mirror/debootstrap/debootstrap_1.0.${minimal_debootstrap}_all.deb
+			sudo dpkg -i debootstrap_1.0.${minimal_debootstrap}_all.deb
+			rm -rf debootstrap_1.0.${minimal_debootstrap}_all.deb || true
+		fi
+	fi
+}
+
 debootstrap_is_installed
 debootstrap_what_version
-
-#if [[ "$test_debootstrap" < "$minimal_debootstrap" ]] ; then
-#if [ "$test_debootstrap" -lt "$minimal_debootstrap" ] ; then
-if [ ! "x$test_debootstrap" = "x$minimal_debootstrap" ] ; then
-	echo "Log: Installing minimal debootstrap version: 1.0.${minimal_debootstrap}..."
-	wget https://rcn-ee.com/mirror/debootstrap/debootstrap_1.0.${minimal_debootstrap}_all.deb
-	sudo dpkg -i debootstrap_1.0.${minimal_debootstrap}_all.deb
-	rm -rf debootstrap_1.0.${minimal_debootstrap}_all.deb || true
-fi
-
+install_debootstrap
